modules/converse-client-server-stdio-demo-local/requirements.txt
fastmcp==0.4.1
boto3==1.37.19
botocore==1.37.19